+++ contracts/access/ContractOwnership.sol
@@ -11,12 +11,10 @@ abstract contract ContractOwnership is ContractOwnershipBase, InterfaceDetection
     using ContractOwnershipStorage for ContractOwnershipStorage.Layout;
 
     /// @notice Initialises the storage with an initial contract owner.
-    /// @notice Sets the ownership storage version to 1.
     /// @notice Marks the following ERC165 interfaces as supported: ERC173.
-    /// @dev Reverts if the ownership storage is already initialized to version `1` or above.
     /// @dev Emits as {OwnershipTransferred} if `initialOwner` is not the zero address.
     /// @param initialOwner the initial contract owner.
     constructor(address initialOwner) {
-        ContractOwnershipStorage.layout().init(initialOwner);
+        ContractOwnershipStorage.layout().constructorInit(initialOwner);
     }
 }