+++ contracts/lifecycle/Pause.sol
@@ -11,11 +11,9 @@ abstract contract Pause is PauseBase, ContractOwnership {
     using PauseStorage for PauseStorage.Layout;
 
     /// @notice Initializes the storage with an initial paused state.
-    /// @notice Sets the pause storage version to `1`.
-    /// @dev Reverts if the pause storage is already initialized to version `1` or above.
     /// @dev Emits a {Paused} event if `isPaused` is true.
     /// @param isPaused the initial pause state.
     constructor(bool isPaused) {
-        PauseStorage.layout().init(isPaused);
+        PauseStorage.layout().constructorInit(isPaused);
     }
 }