+++ contracts/lifecycle/libraries/PauseStorage.sol
@@ -17,16 +17,25 @@ library PauseStorage {
     event Unpaused();
 
     /// @notice Initializes the storage with an initial paused state.
-    /// @notice Sets the pause storage version to `1`.
-    /// @dev Reverts if the pause storage is already initialized to version `1` or above.
+    /// @dev Note: This function should be called ONLY in the constructor of an immutable (non-proxied) contract.
     /// @dev Emits a {Paused} event if `isPaused` is true.
     /// @param isPaused the initial pause state.
-    function init(Layout storage s, bool isPaused) internal {
-        StorageVersion.setVersion(PAUSE_VERSION_SLOT, 1);
+    function constructorInit(Layout storage s, bool isPaused) internal {
         if (isPaused) {
             s.isPaused = true;
             emit Paused();
         }
+    }
+
+    /// @notice Initializes the storage with an initial paused state.
+    /// @notice Sets the pause storage version to `1`.
+    /// @dev Note: This function should be called ONLY in the init function of a proxied contract.
+    /// @dev Reverts if the pause storage is already initialized to version `1` or above.
+    /// @dev Emits a {Paused} event if `isPaused` is true.
+    /// @param isPaused the initial pause state.
+    function proxyInit(Layout storage s, bool isPaused) internal {
+        StorageVersion.setVersion(PAUSE_VERSION_SLOT, 1);
+        s.constructorInit(isPaused);
     }
 
     /// @notice Pauses the contract.