+++ contracts/payment/PayoutWallet.sol
@@ -11,12 +11,10 @@ abstract contract PayoutWallet is PayoutWalletBase, ContractOwnership {
     using PayoutWalletStorage for PayoutWalletStorage.Layout;
 
     /// @notice Initializes the storage with an initial payout wallet.
-    /// @notice Sets the payout wallet storage version to `1`.
-    /// @dev Reverts if the payout wallet storage is already initialized to version `1` or above.
     /// @dev Reverts if `initialPayoutWallet` is the zero address.
     /// @dev Emits a {PayoutWalletSet} event.
     /// @param initialPayoutWallet The initial payout wallet.
     constructor(address payable initialPayoutWallet) {
-        PayoutWalletStorage.layout().init(initialPayoutWallet);
+        PayoutWalletStorage.layout().constructorInit(initialPayoutWallet);
     }
 }