+++ contracts/payment/libraries/PayoutWalletStorage.sol
@@ -16,12 +16,22 @@ library PayoutWalletStorage {
     event PayoutWalletSet(address payoutWallet);
 
     /// @notice Initializes the storage with an initial payout wallet.
+    /// @dev Note: This function should be called ONLY in the constructor of an immutable (non-proxied) contract.
+    /// @dev Reverts if `initialPayoutWallet` is the zero address.
+    /// @dev Emits a {PayoutWalletSet} event.
+    /// @param initialPayoutWallet The initial payout wallet.
+    function constructorInit(Layout storage s, address payable initialPayoutWallet) internal {
+        s.setPayoutWallet(initialPayoutWallet);
+    }
+
+    /// @notice Initializes the storage with an initial payout wallet.
     /// @notice Sets the payout wallet storage version to `1`.
+    /// @dev Note: This function should be called ONLY in the init function of a proxied contract.
     /// @dev Reverts if the payout wallet storage is already initialized to version `1` or above.
     /// @dev Reverts if `initialPayoutWallet` is the zero address.
     /// @dev Emits a {PayoutWalletSet} event.
     /// @param initialPayoutWallet The initial payout wallet.
-    function init(Layout storage s, address payable initialPayoutWallet) internal {
+    function proxyInit(Layout storage s, address payable initialPayoutWallet) internal {
         StorageVersion.setVersion(PAYOUTWALLET_VERSION_SLOT, 1);
         s.setPayoutWallet(initialPayoutWallet);
     }