+++ contracts/proxy/ProxyAdminFacet.sol
@@ -21,7 +21,7 @@ contract ProxyAdminFacet is ProxyAdminBase, ForwarderRegistryContextBase {
     /// @dev Emits an {AdminChanged} event if `initialAdmin` is not the zero address.
     /// @param initialAdmin The initial payout wallet.
     function initProxyAdminStorage(address initialAdmin) external {
-        ProxyAdminStorage.layout().init(initialAdmin);
+        ProxyAdminStorage.layout().proxyInit(initialAdmin);
     }
 
     /// @inheritdoc ForwarderRegistryContextBase