+++ contracts/token/ERC20/ERC20Metadata.sol
@@ -11,11 +11,9 @@ abstract contract ERC20Metadata is ERC20MetadataBase, ContractOwnership {
     using ERC20MetadataStorage for ERC20MetadataStorage.Layout;
 
     /// @notice Initialises the storage with an initial token URI.
-    /// @notice Sets the ERC20Metadata storage version to `1`.
     /// @notice Marks the following ERC165 interface(s) as supported: ERC20Metadata.
-    /// @dev Reverts if the ERC20Metadata storage is already initialized to version `1` or above.
     /// @param uri The token URI.
     constructor(string memory uri) {
-        ERC20MetadataStorage.layout().init(uri);
+        ERC20MetadataStorage.layout().constructorInit(uri);
     }
 }