+++ contracts/token/ERC20/libraries/ERC20PermitStorage.sol
@@ -24,10 +24,18 @@ library ERC20PermitStorage {
     bytes32 internal constant PERMIT_TYPEHASH = keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
 
     /// @notice Initialises the storage.
+    /// @dev Note: This function should be called ONLY in the constructor of an immutable (non-proxied) contract.
+    /// @notice Marks the following ERC165 interface(s) as supported: ERC20Permit.
+    function constructorInit() internal {
+        InterfaceDetectionStorage.layout().setSupportedInterface(type(IERC20Permit).interfaceId, true);
+    }
+
+    /// @notice Initialises the storage.
     /// @notice Sets the ERC20Permit storage version to `1`.
     /// @notice Marks the following ERC165 interface(s) as supported: ERC20Permit.
+    /// @dev Note: This function should be called ONLY in the init function of a proxied contract.
     /// @dev Reverts if the ERC20Permit storage is already initialized to version `1` or above.
-    function init() internal {
+    function proxyInit() internal {
         StorageVersion.setVersion(ERC20PERMIT_VERSION_SLOT, 1);
         InterfaceDetectionStorage.layout().setSupportedInterface(type(IERC20Permit).interfaceId, true);
     }